Edit locaion of the Utils dir to be inside the src dir, as this is more logical
